requirements.txt
# ML Dependencies
# tensorflow-2.4.1-py3-none-any.whl
numpy>=1.23.5
tqdm~=4.66.1
nltk~=3.8.1
gensim~=4.3.2

# Api Dependencies
fastapi~=0.103.0
supabase~=1.0.4

# Local Deps
uvicorn~=0.23.2

python-dotenv~=1.0.0

storage3~=0.5.4
mangum~=0.17.0
tensorflow~=2.14.0
pypdf2~=3.0.1